requirements.txt
aiomysql==0.1.1
alembic==1.16.2
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
Authlib==1.6.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cryptography==45.0.4
dateparser==1.2.1
distro==1.9.0
dnspython==2.7.0
dynaconf==3.2.11
ecdsa==0.19.1
fastapi==0.95.2
fastapi-authlib==0.0.4
fastapi-sa==0.1.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0
google-api-python-client==2.172.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.73.0
grpcio-status==1.71.0
gTTS==2.5.4
h11==0.14.0
httpcore==0.17.3
httplib2==0.22.0
httpx==0.24.1
idna==3.10
inflection==0.5.1
itsdangerous==2.2.0
jiter==0.10.0
Mako==1.3.10
MarkupSafe==3.0.2
motor==3.7.1
oauthlib==3.2.2
openai==1.86.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pycparser==2.22
pydantic==1.10.22
pydantic_core==2.33.2
pygame==2.6.1
PyJWT==2.10.1
pymongo==4.13.1
PyMySQL==1.1.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.5.0
python-multipart==0.0.20
pyttsx3==2.98
pytz==2025.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SpeechRecognition==3.14.3
SQLAlchemy==1.4.46
starlette==0.27.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.4.0
uvicorn==0.21.1
